Split thunks on files. Refactoring
